Allow fast path when setting first ivar on object
When an object is first initialized, ROBJECT_EMBED isn't set. This means
that for brand new objects, ROBJECT_NUMIV(obj) is 0 and
ROBJECT_IV_INDEX_TBL(obj) is NULL.

Previously, this combination meant that the inline cache would never be
initialized when setting an ivar on an object for the first time since
iv_index_tbl was NULL, and if it were it would never be used because
ROBJECT_NUMIV was 0. Both cases always fell through to the generic
rb_ivar_set which would then set the ROBJECT_EMBED flag and initialize
the ivar array.

This commit updates vm_setivar to have another fast path which checks
for the ROBJECT_NUMIV(obj) == 0 case and initializes the ivar array to
Qundef. It also updates the code to set the inline cache to check the
class iv table if the object's iv table hasn't been set.