oh no
Ruby's VALUEs use a tagged pointer scheme, where any of the lowest three
bits being set indicates that the value is "immediate" (not a real
pointer). However the value 0 is also considered a special constant (for
compatibility between Ruby's false and C's false).

This means that to check for a non-heap object required two
comparisons/two conditional branches. I'd wondered for a while if this
could be combined into one with bit twiddling. Turns out it can!

This uses a technique from "Hacker's Delight" to create a word with 1's
in all the positions of training 0's. The following are equivalent:

    (x - 1) & ~x
    (x & -x) - 1
    ~(x | -x)

So in the case of a tagged immediate, we will get a result of either 0
(for a fixnum, with the 1 bit set), 1 (bit 2 is set), or 3 (bit 3 is
set). When we get a 0 the value we get has all bits set, which is -1
when interpreted as a signed integer (the only possible negative value
from this operation). Using this we can perform a signed integer
comparison on the output and anything <= 3 is an special constant!

Is this worthwhile? No idea. There are absolutely downsides both in
complexity and performance.