Spawn new SNT based on grq_cnt > 0
Previously we would spawn new SNT based on the number of total Ractors.
Instead I believe we should do so only when a Ractor is found waiting in
the Global Ready Queue. If the queue is empty, we can assume they're
adequately served by the existing SNTs.

This also avoids a (benign) data-race in the existing code. grq_cnt,
snt_cnt, and dnt_cnt are protected by the scheduler lock. However
vm->ractor.cnt is protected by the VM sync lock, which was not being
held here.