Fix SNT count being off by one
Previously we incremented the snt_cnt local variable when the main
thread was not using MN threading. I believe this was intended to not
count it in the `snt_cnt < vm->ractor.cnt` check, however it also
affected the check for max_cpu. This meant we always had one less SNT
than requested and would deadlock.

[Bug #20146]